fix next logic
the pending slice only contains the pending attendees, not the chair.
the chair itself is a pending attendee though. since this is tricky to
represent in a slice that mutates so often we'll handle this
separately. when the pending slice is empty we'll consider the next user
the chair. upon a new next call we'll then catch the chair and terminate
(reset) the meeting. this puts us into no-meeting state.

should a new attendee appear between the "last" attendee and the chair
this would still get handled properly